v2 push

+++ src/controller/contact.controller.ts
@@ -1,3 +1,4 @@
+import { authenticate } from "../middleware/authenticate-jwt";
 import { prisma } from "../lib/db";
 import { Request, Response } from "express";
 
@@ -13,6 +14,14 @@ export const saveContact = async (req: Request, res: Response) => {
   } = req.body;
 
   try {
+    const authHeader = req.headers['authorization'];
+    const authObj=await authenticate(authHeader)
+    if(!authObj.id){
+      return res.status(401).json({ message: authObj.message });
+    }
+    console.log("v2: saveContact->",new Date());
+
+
     // get the user_id
     const user = await prisma.user.findUnique({ where: { email } });
     if (!user) {
@@ -37,6 +46,14 @@ export const saveContact = async (req: Request, res: Response) => {
 
 export const getAllContacts = async (req: Request, res: Response) => {
   try {
+    const authHeader = req.headers['authorization'];
+    const authObj=await authenticate(authHeader)
+    if(!authObj.id){
+      return res.status(401).json({ message: authObj.message });
+    }
+    console.log("v2: getAllContacts->",new Date());
+
+
     const email = req.query.email as string;
     const user = await prisma.user.findUnique({
       where: {email},
@@ -60,6 +77,13 @@ export const updateContact = async (req: Request, res: Response) => {
   const { name, number, id }:{name:string, number:string, id:string} = req.body;
 
   try {
+    const authHeader = req.headers['authorization'];
+    const authObj=await authenticate(authHeader)
+    if(!authObj.id){
+      return res.status(401).json({ message: authObj.message });
+    }
+    console.log("v2: updateContact->",new Date());
+
     const contact = await prisma.contact.update({
       where: { id },
       data: {
@@ -79,6 +103,13 @@ export const deleteContact = async (req: Request, res: Response) => {
   const { id } = req.body;
 
   try {
+    const authHeader = req.headers['authorization'];
+    const authObj=await authenticate(authHeader)
+    if(!authObj.id){
+      return res.status(401).json({ message: authObj.message });
+    }
+    console.log("v2: deleteContact->",new Date());
+
     await prisma.contact.delete({
       where: { id },
     });
